* miscellaneous OS X porting fixes, mostly just adding casts to   logging calls where a size_t is used as a log argument

+++ applib/APIServer.cc
@@ -203,7 +203,7 @@ APIClient::run()
     ret = readall((char*)&handshake, sizeof(handshake));
     if (ret != sizeof(handshake)) {
         log_err("error reading handshake: (got %d/%u), \"error\" %s",
-                ret, sizeof(handshake), strerror(errno));
+                ret, (u_int)sizeof(handshake), strerror(errno));
         close_session();
         return;
     }
